// sample-project/src/app/protected/modules/samle-module/models/claim-beneficiary-bank-account-details-grrv1.model.ts
export interface ClaimBeneficiaryBankAccountDetailsGRRV1 {
  account_category?: number;
  account_no?: string;
  account_type?: string;
  account_type_code?: string;
  bank_id?: number;
  bank_name?: string;
  country_code?: string;
  country_id?: number;
  country_name?: string;
  currency?: string;
  currency_code?: string;
  foreign_branch_address1?: string;
  foreign_branch_address2?: string;
  foreign_branch_address3?: string;
  foreign_branch_address_city?: string;
  foreign_branch_name?: string;
  ibn_no?: string;
  id?: number;
  local_address1?: string;
  local_address2?: string;
  local_address3?: string;
  local_address_city?: string;
  local_branch_id?: number;
  local_branch_name?: string;
  swift_code?: string;
}