// sample-project/src/app/protected/modules/samle-module/models/beneficiary-bank-account-detail-crv1.model.ts
export interface BeneficiaryBankAccountDetailCRV1 {
  account_name?: string;
  account_no?: string;
  bank_account_category?: number;
  bank_account_detail_id?: number;
  bank_account_type_code?: string;
  bank_id?: number;
  bank_name?: string;
  beneficiary_id?: number;
  claim_note?: string;
  currency_code?: string;
  foreign_bank_branch_address_details?: any;
  foreign_bank_branch_name?: string;
  ibn_no?: string;
  local_bank_branch_id?: number;
  local_bank_branch_name?: string;
  swift_code?: string;
}